SCMNEW-4: refactor for change type creator and last editor to integer

+++ scm-common-data/src/main/java/ir/daneshrefah/scm/common/data/entity/UserTokenDetails.java
@@ -41,7 +41,7 @@ public class UserTokenDetails {
     @Column(name = "ACTIVATION_CODE")
     private String activationCode;
     @Column(name = "CREATED_BY")
-    private Long createdBy;
+    private Integer createdBy;
     @Column(name = "MODIFIED_BY")
     private Integer modifiedBy;
     @Column(name = "CREATION_DATE")
