+++ scm-uaa-common/src/main/java/ir/daneshrefah/scm/uaa/common/utils/AuthenticationUtils.java
@@ -214,7 +214,7 @@ public class AuthenticationUtils {
         IpAddressMatcher ipAddressMatcher = new IpAddressMatcher(ipAddress);
         return ipAddressMatcher.matches(matchesIpAddress);
     }
-    public static Long getLoggedInUserId() {
+    public static Integer getLoggedInUserId() {
         User loggedInUser = AuthenticationUtils.getLoggedInUser();
         assert loggedInUser != null;
         return loggedInUser.getPerson().getId();
